feat: add webSocketServer
