+++ .eslintrc
@@ -1,0 +1,11 @@
+{
+  "parser": " @typescript-eslint/parser",
+  "extends": [ "plugin:  @typescript-eslint/recommended" ],
+  "parserOptions": {
+    "ecmaVersion": 2018,
+    "sourceType": "module"
+  },
+  "rules": {
+
+  }
+}
