+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["./"],
+  "ext": ".ts, .js",
+  "ignore": [],
+  "exec": "npx ts-node ./index.ts"
+}
