+++ package.json
@@ -2,24 +2,25 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
+    "main": "index.ts",
     "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "node index.ts",
+        "start:dev": "npx nodemon",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Davydau Mikalai",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.20",
-        "ts-node": "^10.9.1",
         "@types/node": "^18.11.17",
+        "@types/ws": "^8.5.4",
         "@typescript-eslint/eslint-plugin": "^5.47.0",
         "@typescript-eslint/parser": "^5.47.0",
         "eslint": "^8.30.0",
+        "nodemon": "^2.0.20",
         "prettier": "^2.8.1",
+        "ts-node": "^10.9.1",
         "typescript": "^4.9.4"
     },
     "dependencies": {
